add user creation

+++ backend/userguard/src/main/resources/db/migration/V1__Create_Users_Table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    id BIGINT PRIMARY KEY,
+    id serial PRIMARY KEY,
     username VARCHAR(255) UNIQUE NOT NULL,
     password VARCHAR(255) NOT NULL,
     email VARCHAR(255) UNIQUE NOT NULL,
